refacto: add buttons namespace
