+++ src/apis.ts
@@ -1,8 +1,7 @@
 import PubSub from '@/api/pubSub'
 import Chat from '@/api/chat'
 import Streams from '@/api/streams'
-import RegisterShareButton from '@/api/registerShareButton'
-import RegisterStageButton from '@/api/registerStageButton'
+import * as Buttons from '@/api/buttons'
 import Theme from '@/api/theme'
 import NotificationCenter from '@/api/notificationCenter'
 import Storage from '@/api/storage'
@@ -14,16 +13,15 @@ import register from '@/api/register'
 
 export {
   Configuration,
+  Buttons,
+  Chat,
+  NotificationCenter,
+  Modal,
   PubSub,
   Streams,
   Storage,
-  RegisterShareButton,
-  RegisterStageButton,
-  NotificationCenter,
-  Modal,
-  Users,
-  Chat,
   Theme,
+  Users,
   When,
   register,
 }
