+++ src/types/button.ts
@@ -1,19 +1,16 @@
-export interface ShareButtonOptions {
+export interface DefaultButtonOptions {
     label: string,
     icon?: string,
     imageSource?: string,
-    onClick: () => unknown
+    onClick?: (payload?: Record<string, unknown>) => unknown
 }
 
-export type StageButton = {
+export type StageButtonWrapper = {
     remove: () => void
 }
 
-export interface StageButtonOptions {
-    label?: string,
+export interface StageButtonOptions extends DefaultButtonOptions {
     tooltip?: string,
-    icon?: string,
-    imageSource?: string,
     dropdownActions?: Array<{ name?: string, label: string, imageSource?: string }>,
     dropdownActionsTextClasses?: string,
     iframe?: {
@@ -22,6 +19,4 @@ export interface StageButtonOptions {
       width: number,
       height: number,
       onMessage?: (payload?: Record<string, unknown>) => unknown,
-    },
-    onClick?: (payload?: Record<string, unknown>) => unknown
-}
+    },}
